+++ features/admin.registration-flow-builder.v1/components/element-property-panel/component-property-factory.tsx
