+++ features/admin.registration-flow-builder.v1/components/element-property-panel/element-properties.tsx
