+++ features/admin.registration-flow-builder.v1/components/element-property-panel/extended-properties/field-extended-properties.tsx
