+++ features/admin.registration-flow-builder.v1/components/element-property-panel/widgets/attribute-collector-properties.scss
